slight re-org for session reaper and workspaces
